backend/requirements.txt
flask
flask-cors
# flask-mysql
mysql-connector-python
